Allow for specific integer conversion in database.
By default, all values in the database are strings. This is problematic
when we come to weights, as they are treated like integers in the
WeightedDatabase code, but are actually stored as string.

To side-step implicit conversion, instead use explicit conversion by an
"%int <key>" signifier in the database specification.