Properly deal with classes of varying inheritance in documentation.
Previously the documentation assumed that all classes grouped together
had the same inheritance schema. Interestingly, this was the case:
Shape, ShapeCollection, etc, all inherit directly from object; likewise
Coord and so-on; likewise Box and Column inherit from Shape.

However, adding a section that involved multiple classes that inherit
from differing classes (AutoSize and AutoDimension, for instance)
resulted in the tree being:

[
 (<type 'int'>, (<type 'object'>,)), [
 	(<class 'coord.AutoDimension'>, (<type 'int'>,))
 ],
 (<class 'coord.Size'>, (<class 'coord.Coord'>,)), [
	(<class 'coord.AutoSize'>, (<class 'coord.Size'>,))
 ]
]

The code previous only investigated tree[1], assuming that the classes
all inherited from tree[0]. As noted above, this was the case for
classes inheriting from object or a common ancestor, but not the case
for this instance.

Instead, we iterate over every second member of the tree (ie, all
evens), assuming that every first member of the tree is a base type.