Update ShapeCollection. No sorting, AutoShape.
We don't sort automatically any more, so don't mention it in the
documentation! Likewise, use an AutoShape as the basis for combination
rather than trying to find the largest Shape contained within.

Stubs out the width, height and size parameters until I can think up a
saner way to implement them (without actually performin a draw and then
returning the size of the resultant canvas, which is cheap and nasty and
not necessarily cheap).