Make AutoShape::normalise function properly by wrapping.
As it calls unbound Shape::normalise(AutoShape) instead of a member
function of AutoShape, it doesn't take into consideration the
AutoDimensions returned whenever width, height, etc, are called.

This caused problems because normalise performs sanity checking to
ensure that we aren't trying to normalise a shap to smaller than the
shape is.

Instead of re-implementing normalise specifically for AutoShape, I've
instead implemented a wrapper function: it hot-swaps the relevant
functions that are expected to be in place into place, executes the
function, and then reverses the hot-swapping.