New ShapeCollection methods: extend, offset.
The first allows simple extension of one collection by another (allowing
for multiplication; I'd write an __add__, etc, suite of functions for
the class if I thought they'd ever be needed, but I think using
list-like functions here is cleaner and more easily understood).

The second iterates over the collection of shapes and performs an offset
with the relevant value. This is useful when you want to shift a
selection of shapes down, up, left, right, etc.