Fix adjoin, atop, etc's uses of ShapeCollection.
Now you can pass a collection as the first paramter and a shape as the
second; the shape will be offset based on the current contents of the
collection and then appended.

You can also pass two collections: the second will be offset as
necessary, and then the two will be added together (or rather, each
shape in the second will be added to the first).