Don't "from <x> import *": it causes issues with things being in the wrong modules.
ie, it was possible to "import shape" and use "shape.ShapeCollection"
because the shape library had "from collection import *". This is bad
and stuffs things into namespaces they shouldn't be in.

Instead, import the module and use that. Thus there are a couple of
instances where "coord" as a variable has to be changed to "c";
collection as a variable to "collect", etc. Hopefully I've caught them
all.