Move colour and feature modules into the library.
These are classes and abstracts, for the most part, rather than actual
information. I'd rather that the interface-specific things, such as
lists of features (which are variable and subject to change) were in
that package, while common or standardised things (such as colours,
which are not subject to change with much regularity) are in the library
package.