Define a rough interface as to what functions should be provided.
Specifically, a "common" interface that can be used as an abstraction
layer. I've tried to keep it as simple as possible: we don't need to
worry about what's *on* the screen at the minute, or such things as
cursors, etc, because each of these will be handled agnostically.

It's not necessarily the sanest option, but we'll see what happens.