Some work on cleaning up the win32 console module.
Instead of attempting to force 80x25, which doesn't work, instead try to
select the current screen width and just force the buffer scrollback to
25 lines.

This results in a rather pleasant screen regardless of configuration.
However, it would still be good if we could actually redraw forcefully
the screen at the desired screen width and height.