curses: Curses interface colours now works completely!
Well, as complete as I've tested it.

The abstraction between screen characters and screen colours needs to be
written, but from here it should be a relatively simple issue. There
also needs to be an event loop...

Being able to get identical results out of the win32 interface is the
next step.