In test_walk, display the id of the room you're walking through.
For some reason, the corridor is still considered as part of most
of the adjacent rooms. What might work is having corridor coords
simply not included in any rooms, so that, the other way around,
we could know that if a position lies within the manor boundaries
(i.e. there are rooms all around it) but is not within a room
itself, it must be a corridor.

(Unfortunately, that rule of thumb won't work anymore with the
O-layout.)

On the upside, if it's that easy to get the room id, matching
room types and properties to ids shouldn't be hard either.