Add doors to the manor.
The algorithm uses the fact that walls are mostly defined by overlaps
between room/corridor coordinates. For each corridor, picks a random
wall position for each of the adjacent rooms, with the caveat that
the coordinate is only shared by 1 other room. The latter means that
some rooms don't get doors because the small section overlapping with
the corridor also happens to overlap with other rooms. Still, this
looks quite nice.

I've also worked out how to directly draw onto the canvas, so doors
get drawn right in. Currently hardcoded '+', but that can be changed.

Still contains lots of debugging output, as there's still a lot of
room for improvement.