Plug a hole in the wall, and simplify some debugging output.
Whatever a corridor may pretend, if it reaches the manor boundary, we
know there's a wall there. :P

I've found out what causes the big holes where the legs are attached:
Tnstead of changing the room's size or position (which is apparently
not possible in the current implementation), the corridor is simply
drawn on top of it, so it appears correctly on the canvas, but for
the feature grid, the room's wall coincides with the corridor's one,
so no additional wall is drawn.