Don't scroll if movement is blocked.
Scrolling is still far from optimal. Ideally, the screen should show
as much of the manor as can fit, while trying to center the visible
section on the player.

Also, replace the "you can't move there" X markers with a proper message.
Yay, this is starting to feel like an actual game!