Add a RoomProps object for storing room properties such as name/type.
RoomProps currently inherits from Room because it's possible that
some of the rooms' positions and size need to be updated. So far I've
been able to work around this restriction, but that might not always
be the case.

The manor keeps a list of these with RoomProps indices matching the
room ones. Eventually, the former might replace the latter.