Split test_mockup's main method into different Game module methods.
Game currently resides in test_mockup.py itself but will eventually
get a file of its own. Should be more readable and easier to maintain.

Separation of the drawing methods for viewport and messages allows
us avoiding unnecessary drawing operations. This is currently unused
because the running checks are a bit clunky.