Start work on the "screen" object and interface.
It's nowhere near complete, but at the very least there is now some
sort of abstraction between the actual console drawing interface,
and how we print to it.

Eventually it needs to be able to handle multiple views, but initially
the main thing that I'm working on is a scrolling "message" region,
a "fixed" region for displaying the main information about the game,
and a viewport region for actually displaying the manor.