Given a list of (id, name) tuples, assign bed rooms.
Still doesn't make use of the suspect list (though we're getting
closer), just tested with a dummy list for now.

Bedrooms may never be a passage room (though there's currently
nothing to prevent requiring to walk through one to get there) and,
if possible, should use windows.

The manor currently requires 5 non-bedrooms, plus the entrance hall.
This limit seems a bit low, but only really matters for the smaller
manor layouts, where it needs to be low.

Once the utility wing has been filled, bedrooms are then put in the
remaining wings (if any) before the central area is considered.