New test file: python -m builder.test.test_map
Outputs an annotated map of the manor with numbered rooms and a legend,
that can easily be piped into a file. Accepts the manor layout
('L', 'U', 'H') as parameter; default is the base_builder.

Also contains all the debug garbage at the top. Scroll down to get to
the interesting stuff.