Further monkey-patching: __init__ methods, correct signature for classes.
While the autodoc extension knows that it needs to fetch the argspec for
a class definition from the class's __init__ method, if it exists, it
doesn't emit an event that allows the user to alter the signature --
like it does with ever other signature.

This means that if you manually tell it to document the __init__ method,
and you, like us, have a function that modifies the signature of
derived-functions, you will get a "correct" signature for that method,
and the actual signature (the "incorrect" one) for the class itself.

Annoying!