Hook autotravel into mainloop.py.
Pulls up a menu of rooms to choose from, then tries to find a path to
the centre of the room. (This may not be possible due to a large
feature taking up that position, but it's only a temporary solution
anyway.)

For some reason, the intermediary steps do not get drawn, or rather
the screen doesn't get updated until after you arrive. Seeing how my
computer was wheezing when I was only trying to reach the next room,
I've disabled the redrawing attempt for now. It's possible we may not
even want to redraw the screen while traveling, but I'd at least like
to know _why_ it doesn't work.