Allow an alternate pathfinding target condition as a function parameter.
If both a target coordinate and a target function are passed in,
pathfinding stops when the first of them is reached.

Make use of this by stopping mainloop.py pathfinding once the target
room has been reached (rather than in the middle of the room).

Other applications could be: finding the nearest staircase, or stopping
once a given distance has been reached.